% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idw_inter.R
\name{idw_inter}
\alias{idw_inter}
\title{Get natural earth world state (admin level 1) polygons}
\usage{
idw_inter(dat, sta, cntr, coun_cd, alt, reg)
}
\arguments{
\item{dat}{a character vector of MODIS Sensor name "Terra, Aqua".}

\item{sta}{a character vector of USERNAME of Earthdata Login.}

\item{cntr}{a character vector of PASSWORD of Earthdata Login.}

\item{coun_cd}{a character vector, format YYYY.MM.DD, of Start date}

\item{alt}{a character vector, format YYYY.MM.DD, of End date}

\item{reg}{a character vector, name of the Region of the Country}
}
\description{
returns GTiff file of Land Surface Temperature for specified region or country
}
\examples{

file_path <- system.file("extdata", "temp.csv", package = "coriM")
file <- readr::read_csv(file_path)

idw_inter(dat = file, sta = TRUE, cntr="peru",coun_cd = 'PE',
 alt = TRUE, reg = "Puno")


}
