% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load Sample Data
Function in case you want to observe the type of data to be interpolated
with the IDW method.}
\usage{
loadData()
}
\description{
This function loads the sample data from the package's extdata directory.
}
