% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_discrete_cbar.R
\name{plot_discrete_cbar}
\alias{plot_discrete_cbar}
\title{This function generates data interpolation using the IDW method. The monthly
data acquired was obtained from meteorological stations with maximum and
minimum temperature information for the year 2010 in March, for the whole
region of Puno.}
\usage{
plot_discrete_cbar(
  breaks,
  palette = "Greys",
  colors = RColorBrewer::brewer.pal(length(breaks) - 1, palette),
  direction = 1,
  spacing = "natural",
  border_color = NA,
  legend_title = NULL,
  legend_direction = "horizontal",
  font_size = 5,
  expand_size = 1,
  spacing_scaling = 1,
  width = 0.1,
  triangle_size = 0.1
)
}
\description{
#'
}
