% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selCountry.R
\name{selCountry}
\alias{selCountry}
\title{Function to check all the regions in the console and selected}
\usage{
selCountry(cntry)
}
\arguments{
\item{cntry}{character vector country name}
}
\description{
Function to check all the regions in the console and selected
}
\examples{

selCountry(cntry = "peru")

}
