% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idw_inter.R
\name{idw_inter}
\alias{idw_inter}
\title{This function generates data interpolation using the IDW method. The monthly
data acquired was obtained from meteorological stations with maximum and
minimum temperature information for the year 2010 in March, for the whole
region of Puno.}
\usage{
idw_inter(dat, sta, cntr, coun_cd, alt, reg)
}
\arguments{
\item{dat}{the path of the CSV file to read.}

\item{sta}{a LOGIC character if you need Country or Regional Boundary.}

\item{cntr}{a character vector of country name}

\item{coun_cd}{a character vector of ISO CODE of the country}

\item{alt}{a LOGIC character, if you need the Altitude or not}

\item{reg}{a character vector, name of the Region of the Country}
}
\description{
This function generates data interpolation using the IDW method. The monthly
data acquired was obtained from meteorological stations with maximum and
minimum temperature information for the year 2010 in March, for the whole
region of Puno.
}
\examples{

Load the data example

file_path <- system.file("extdata", "temp.csv", package = "coriM")
file <- readr::read_csv(file_path, show_col_types = FALSE)

Run the IDW function

idw_inter(dat = file, sta = TRUE, cntr="peru",coun_cd = 'PE',
 alt = TRUE, reg = "Puno")


}
