% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selCountry.R
\name{selCountry}
\alias{selCountry}
\title{This function will show me in console, all the regions of the country
selected for the calculation in case someone wants to get information of
a more precise area.}
\usage{
selCountry(cntry)
}
\arguments{
\item{cntry}{character vector country name}
}
\description{
This function will show me in console, all the regions of the country
selected for the calculation in case someone wants to get information of
a more precise area.
}
\examples{

selCountry(cntry = "peru")

}
